separate DSL from view helper
