+++ app/components/primer/view_components.rb
@@ -14,6 +14,7 @@ require "octicons_helper/helper"
 
 require "primer/class_name_helper"
 require "primer/classify"
+require "primer/dsl/view_helper"
 require "primer/fetch_or_fallback_helper"
 require "primer/join_style_arguments_helper"
 require "primer/view_helper"
