+++ lib/primer/view_helper.rb
@@ -1,27 +1,11 @@
 # frozen_string_literal: true
 
-require 'active_support/concern'
-
 module Primer
   module ViewHelper
     extend ActiveSupport::Concern
 
-    class ViewHelperAlreadyDefined < StandardError; end
-    class ViewHelperNotFound < StandardError; end
-
-    included do
-      class_attribute :helpers, instance_writer: false, default: {}
-    end
-
-    class_methods do
-      def view_helper(name)
-        raise ViewHelperAlreadyDefined if helpers[name].present?
-        helpers[name] = self
-      end
-    end
-
     def primer(name, **component_args, &block)
-      component = helpers[name]
+      component = Primer::Component.helpers[name]
 
       raise ViewHelperNotFound unless component.present?
